+++ src/lib/repl/actions/mouse_slider.ts
@@ -13,32 +13,53 @@ export interface IMouseSliderOptions {
 }
 
 export function mouse_slider(element: HTMLElement, options: IMouseSliderOptions) {
-    // TODO: Touch support?
-
     let {horizontal = false, on_move, on_state, target = element} = options;
     let grabbing = false;
 
-    function on_mousedown(event: MouseEvent) {
-        const {button} = event;
+    function get_touch_position(event: MouseEvent | TouchEvent): [number, number] {
+        if ("touches" in event) {
+            const [touch] = event.touches;
 
-        if (button === 0) {
+            return [touch.clientX, touch.clientY];
+        }
+
+        return [event.clientX, event.clientY];
+    }
+
+    function is_touching(event: MouseEvent | TouchEvent): boolean {
+        if ("button" in event) {
+            return event.button === 0;
+        } else if ("touches" in event) {
+            // TODO: What would be the proper check for this? And to
+            // detect long presses or something?
+            return event.touches.length > 0;
+        }
+
+        return false;
+    }
+
+    function on_inputdown(event: MouseEvent | TouchEvent) {
+        if (is_touching(event)) {
             grabbing = true;
             if (on_state) on_state(true);
         }
     }
 
-    function on_mouseup(event: MouseEvent) {
+    function on_inputup(event: Event) {
         if (grabbing) {
             grabbing = false;
             if (on_state) on_state(false);
         }
     }
 
-    function on_mousemove(event: MouseEvent) {
+    function on_inputmove(event: MouseEvent | TouchEvent) {
         if (!grabbing) return;
 
+        event.preventDefault();
+
         const rect = element.getBoundingClientRect();
-        const cursor = horizontal ? event.clientX : event.clientY;
+        const [client_x, client_y] = get_touch_position(event);
+        const cursor = horizontal ? client_x : client_y;
 
         const minimum = horizontal ? rect.left : rect.top;
         const maximum = horizontal ? rect.right : rect.bottom;
@@ -49,23 +70,27 @@ export function mouse_slider(element: HTMLElement, options: IMouseSliderOptions)
         on_move(position / size);
     }
 
-    element.addEventListener("mouseup", on_mouseup);
-    element.addEventListener("mousemove", on_mousemove);
-    target.addEventListener("mousedown", on_mousedown);
+    element.addEventListener("mouseup", on_inputup);
+    element.addEventListener("touchend", on_inputup);
+    element.addEventListener("mousemove", on_inputmove);
+    element.addEventListener("touchmove", on_inputmove);
+
+    target.addEventListener("mousedown", on_inputdown);
+    target.addEventListener("touchstart", on_inputdown);
 
     return {
         update(options: IMouseSliderOptions) {
-            target.removeEventListener("mousedown", on_mousedown);
+            target.removeEventListener("mousedown", on_inputdown);
 
             ({horizontal = false, on_move, on_state, target = element} = options);
 
-            target.addEventListener("mousedown", on_mousedown);
+            target.addEventListener("mousedown", on_inputdown);
         },
 
         destroy() {
-            element.removeEventListener("mouseup", on_mouseup);
-            element.removeEventListener("mousemove", on_mousemove);
-            target.removeEventListener("mousedown", on_mousedown);
+            element.removeEventListener("mouseup", on_inputup);
+            element.removeEventListener("mousemove", on_inputmove);
+            target.removeEventListener("mousedown", on_inputdown);
         },
     };
 }